requirements.txt
numpy
torch
openai-whisper